crypto-airflow-pipeline/requirements.txt
pandas
requests
matplotlib
psycopg2-binary
apache-airflow==2.7.2
